Test that opening a directory gives the right error.
Send the right supported-block-vector and supported-open-block-vector
in supported2.

Avoid overflow in allocation-size calculation.

Error on various desired-access bits.

+++ stat.c
@@ -60,7 +60,8 @@ void stat_to_attrs(struct allocator *a,
   default: attrs->type = SSH_FILEXFER_TYPE_SPECIAL; break;
   }
   attrs->size = sb->st_size;
-  attrs->allocation_size = sb->st_blksize * sb->st_blocks;
+  /* The cast ensures we don't do the multiply in a small word and overflow */
+  attrs->allocation_size = (uint64_t)sb->st_blksize * sb->st_blocks;
   /* Only look up owner/group info if wanted */
   if((flags & SSH_FILEXFER_ATTR_OWNERGROUP)
      && (attrs->owner = uid2name(a, sb->st_uid))
